backend/requirements.txt
# AnIntelligentAI Backend - Core Dependencies

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis (Short-term Memory)
redis==5.0.1

# Qdrant (Vector Database)
qdrant-client==1.7.0

# LLM Providers
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2

# Utilities
tenacity==8.2.3
python-dotenv==1.0.0

# Document Processing (RAG)
PyPDF2==3.0.1
python-docx==1.1.0

# Security & Encryption
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging & Monitoring
python-json-logger==2.0.7

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
numpy==1.26.3

# Production Server
gunicorn==21.2.0

# Testing (Optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development (Optional)
black==23.12.1
flake8==7.0.0
mypy==1.8.0
